SWL-40 Converted MainContainer to use Redux Hooks, removed MainContainer2, moved connect() of the original MainContainer to MainContainer.tsx-backup
Changes to be committed:
	modified:   src/client/components/App.tsx
	renamed:    src/client/components/main/MainContainer2.tsx -> src/client/components/main/MainContainer.backup
	modified:   src/client/components/main/MainContainer.tsx
	deleted:    src/client/components/main/MainContainer3.tsx
