+++ src/client/components/App.tsx
@@ -11,9 +11,7 @@ import collectionsController from '../controllers/collectionsController';
 // import UpdatePopUpContainer from './legacy-components/UpdatePopUpContainer';
 import HistoryOrWorkspaceContainer from './workspace/HistoryOrWorkspaceContainer';
 import NavBarContainer from './navbar/NavBarContainer';
-// import MainContainer from './main/MainContainer';
-
-import MainContainer2 from './main/MainContainer2';
+import MainContainer from './main/MainContainer';
 
 // Types
 import { WindowExt } from '../../types';
@@ -64,8 +62,7 @@ const App = () => {
 
           {/* Main container, Contains the composer and response panes. */}
           <ErrorBoundary>
-            {/* <MainContainer currentWorkspaceId={currentWorkspaceId} /> */}
-            <MainContainer2 />
+            <MainContainer currentWorkspaceId={currentWorkspaceId} />
           </ErrorBoundary>
         </Split>
